requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
python-dotenv==1.0.1

openai==1.72.0

youtube-transcript-api==1.2.3
